+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Added messageId to MessageDetail dialog
 - added custom titlebar menu for webxdc (on linux and windows)
+- add open keybindings dialog to window titlebar menu
 
 ### Fixed
 - "New contact" button hidden if contact already exists (#2646)